[CWS] fix TestMain not called
